# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Oblivion API
  version: "1.0.0"
  description: |
    REST API for **Oblivion**, a face / genetic-syndrome recognition backend.

    The system is split into 7 independent microservices. They sit behind an
    nginx **API gateway** at `http://localhost:8080`, which routes by path
    prefix to the right service — so this whole API is one origin and
    "Try it out" works without switching ports. Endpoints are grouped by the
    **tag** (= service); the direct per-service ports are noted on each tag for
    when you want to bypass the gateway.

    ## Conventions
    - Every response is wrapped in an envelope:
      `{ "status": "ok" | "error", "data": <payload> | null, "error": <string> | null }`.
    - Authenticated endpoints expect `Authorization: Bearer <access_token>`,
      a RS256 JWT issued by the auth service.
    - Roles: `doctor`, `lab`, `admin`. Lab endpoints require a `lab` token.

    ## Typical flow
    1. `POST /auth/register` then `POST /auth/login` → access token.
    2. `POST /cases` → returns a presigned S3 `upload_url`.
    3. `PUT` the image bytes to `upload_url` (MinIO/S3, no auth).
    4. `POST /cases/{id}/confirm-upload` → triggers async analysis (Kafka).
    5. Poll `GET /cases/{id}` until `status = completed`; read `diagnoses`.
servers:
  - url: http://localhost:8080
    description: API gateway (nginx) — routes by path prefix to all services
...
paths:
...
  /cases/{id}/confirm-upload:
    parameters: [{ $ref: '#/components/parameters/IdPath' }]
    post:
      tags: [cases]
      summary: Confirm image upload and start analysis
      description: >
        Verifies the image exists in S3, moves the case to `processing`, and
        publishes an analysis request onto Kafka.
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/OkEmpty' }
        '404':
          description: Case or uploaded image not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
        '409':
          description: Case is not in the pending state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }

  # ── labs (3003) ────────────────────────────────────────────────────────────
  /labs/batch:
    post:
      tags: [labs]
      summary: Submit a batch of images for analysis
      description: Checks the lab's quota with the billing service, then enqueues each image.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [image_keys]
              properties:
                image_keys:
                  type: array
                  items: { type: string }
                  example: ["batch1/a.jpg", "batch1/b.jpg"]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties: { data: { $ref: '#/components/schemas/BatchWithItems' } }
        '402':
          description: Quota exceeded / no active subscription
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
        '401': { $ref: '#/components/responses/Error' }
  /labs/results/{batch_id}:
    parameters:
      - name: batch_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [labs]
      summary: Get batch status and per-item results
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties: { data: { $ref: '#/components/schemas/BatchWithItems' } }
        '404': { $ref: '#/components/responses/Error' }
  /labs/usage:
    get:
      tags: [labs]
      summary: Current month usage for the calling lab
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK (data is null until the lab has a subscription)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data:
                        oneOf:
                          - { $ref: '#/components/schemas/UsageInfo' }
                          - { type: 'null' }
  /labs/tiers:
    get:
      tags: [labs]
      summary: List available subscription tiers
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Plan' }

  # ── recognition (3004) ─────────────────────────────────────────────────────
...
  /notifications/{id}/read:
    parameters: [{ $ref: '#/components/parameters/IdPath' }]
    patch:
      tags: [notifier]
      summary: Mark a notification as read
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties: { data: { $ref: '#/components/schemas/Notification' } }
        '404': { $ref: '#/components/responses/Error' }

  # ── reference (3006) ───────────────────────────────────────────────────────
...
  /reference/hpo/search:
    get:
      tags: [reference]
      summary: Search HPO (Human Phenotype Ontology) terms
      security: [{ bearerAuth: [] }]
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
          example: micro
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/HpoTerm' }

  # ── billing (3007) ─────────────────────────────────────────────────────────
...
  /billing/usage/{lab_id}:
    parameters:
      - name: lab_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [billing]
      summary: Get a lab's current usage
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties: { data: { $ref: '#/components/schemas/UsageInfo' } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
...
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
    OkEmpty:
      description: Success envelope with null data
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
  schemas:
    Envelope:
      type: object
      description: Standard response wrapper used by every endpoint.
      required: [status]
      properties:
        status: { type: string, enum: [ok, error], example: ok }
        data: { description: Payload (shape varies by endpoint), nullable: true }
        error: { type: string, nullable: true, example: null }
...
    CaseStatus:
      type: string
      enum: [pending, processing, completed, failed]
...
    Batch:
      type: object
      properties:
        id: { type: string, format: uuid }
        lab_id: { type: string, format: uuid }
        status: { $ref: '#/components/schemas/CaseStatus' }
        total_items: { type: integer }
        processed_items: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    BatchItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        batch_id: { type: string, format: uuid }
        image_key: { type: string }
        case_id: { type: string, format: uuid, nullable: true }
        status: { $ref: '#/components/schemas/CaseStatus' }
        created_at: { type: string, format: date-time }
    BatchWithItems:
      type: object
      properties:
        batch: { $ref: '#/components/schemas/Batch' }
        items:
          type: array
          items: { $ref: '#/components/schemas/BatchItem' }
...
    Notification:
      type: object
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        type: { type: string, example: analysis_complete }
        title: { type: string }
        body: { type: string, nullable: true }
        read: { type: boolean }
        metadata: { type: object, additionalProperties: true, nullable: true }
        created_at: { type: string, format: date-time }
...
    HpoTerm:
      type: object
      properties:
        id: { type: string, format: uuid }
        hpo_id: { type: string, example: "HP:0000252" }
        name: { type: string }
        definition: { type: string, nullable: true }
...
    Plan:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: bronze }
        monthly_limit: { type: integer, example: 40 }
        overage_price_cents: { type: integer, example: 5000 }
        created_at: { type: string, format: date-time }
...
    UsageInfo:
      type: object
      properties:
        lab_id: { type: string, format: uuid }
        month: { type: string, format: date, example: "2026-05-01" }
        plan_name: { type: string, example: bronze }
        monthly_limit: { type: integer }
        current_usage: { type: integer }
        remaining: { type: integer }
...